% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPSC2-package.R
\docType{data}
\name{family_cn}
\alias{family_cn}
\title{Plant Family Names}
\format{
A data frame with 637 observations on the following 2 variables.
\describe{
\item{FAMILY_CN}{a character vector representing the Chinese names of the family}
\item{FAMILY}{a character vector representing the family} }
}
\description{
Plant Family Names
}
\examples{

data(family_cn)
## maybe str(family_cn) ; plot(family_cn) ...

}
\keyword{datasets}
